Improved fluid rendering and indicating in GUIs
- StoredFluidWidget now renders it's stored fluid according to the
fluid's still texture defined in the fluid's FluidType.
- CreativeFluidSource's name now changes to say what fluid it's holding
if it's holding a fluid besides Empty.
- CreativeFluidStorer's name now changes if it's holding a fluid besides
Empty to say what fluid it's holding, and to what degree it's currently
filled.
- Added fluid translations.
- Added convenient possibly unsafe method of transmitting the fluid
that's stored in a block entity from server to a client menu so it can
be rendered in any GUIs.
- Fixed an int literal using 0x instead of 0b in the
ExtractorBlockEntity class.
- A bit out of theme: added soul sap extraction from soul afflicted
rubber wood tree blocks to match latex extraction from rubber tree
blocks.
- Probably added bugs.